move unused method to tests
